+++ Flow/src/utopia/flow/datastructure/immutable/Tree.scala
@@ -1,11 +1,14 @@
 package utopia.flow.datastructure.immutable
 
+import utopia.flow.util.EqualsFunction
+
 /**
  * This TreeNode implementation is immutable and safe to reference from multiple places
  * @author Mikko Hilpinen
  * @since 4.11.2016
  */
 case class Tree[A](override val content: A, override val children: Vector[Tree[A]] = Vector())
+                  (implicit equals: EqualsFunction[A] = EqualsFunction.default)
     extends TreeLike[A, Tree[A]]
 {
     // IMPLEMENTED    ---------------
@@ -13,6 +16,8 @@ case class Tree[A](override val content: A, override val children: Vector[Tree[A
     override def repr = this
     
     override protected def createCopy(content: A, children: Vector[Tree[A]]) = Tree(content, children)
+    
+    override def containsDirect(content: A) = equals(this.content, content)
     
     override protected def newNode(content: A) = Tree(content)
     
@@ -25,7 +30,8 @@ case class Tree[A](override val content: A, override val children: Vector[Tree[A
       * @tparam B Target content type
       * @return A mapped version of this tree
       */
-    def map[B](f: A => B): Tree[B] = Tree(f(content), children.map { _.map(f) })
+    def map[B](f: A => B)(implicit equals: EqualsFunction[B]): Tree[B] =
+        Tree(f(content), children.map { _.map(f) })(equals)
     
     /**
       * Maps the content of all nodes in this tree. May map to None where the node is removed.
@@ -33,5 +39,6 @@ case class Tree[A](override val content: A, override val children: Vector[Tree[A
       * @tparam B Target content type
       * @return A mapped version of this tree. None if the content of this node mapped to None.
       */
-    def flatMap[B](f: A => Option[B]): Option[Tree[B]] = f(content).map { c => Tree(c, children.flatMap { _.flatMap(f) }) }
+    def flatMap[B](f: A => Option[B])(implicit equals: EqualsFunction[B]): Option[Tree[B]] =
+        f(content).map { c => Tree(c, children.flatMap { _.flatMap(f) })(equals) }
 }