+++ Flow/src/utopia/flow/parse/NamespacedString.scala
@@ -1,10 +1,16 @@
 package utopia.flow.parse
 
-import utopia.flow.util.Equatable
+import utopia.flow.util.{EqualsFunction, Equatable}
 import utopia.flow.util.StringExtensions._
 
 object NamespacedString
 {
+	/**
+	  * An equality function that checks whether the two strings match each other. The check is case-insensitive and
+	  * an empty namespace is considered to match with every other namespace.
+	  */
+	implicit val approxEquals: EqualsFunction[NamespacedString] = _ ~== _
+	
 	/**
 	  * @param local Local part of this string
 	  * @param namespace Namespace applied to the local string
@@ -26,6 +32,10 @@ case class NamespacedString(local: String)(implicit val namespace: Namespace) ex
 	  * @return Whether this string has a namespace definition
 	  */
 	def hasNamespace = namespace.nonEmpty
+	/**
+	  * @return Whether this string doesn't have an associated namespace
+	  */
+	def hasNoNamespace = namespace.isEmpty
 	
 	/**
 	  * @return Namespace applying to this string. None if empty namespace is applied.
@@ -37,11 +47,12 @@ case class NamespacedString(local: String)(implicit val namespace: Namespace) ex
 	
 	override def properties = Vector(local, namespace)
 	
-	override def toString = s"$namespace:$local"
+	override def toString = if (hasNamespace) s"$namespace:$local" else local
 	
 	
 	// OTHER    ---------------------------------
 	
 	def ==(localString: String) = local == localString
 	def ~==(localString: String) = local ~== localString
+	def ~==(other: NamespacedString) = (local ~== other.local) && (namespace ~== other.namespace)
 }