+++ Flow/test/utopia/flow/test/parse/XmlReadTest.scala
@@ -1,7 +1,9 @@
 package utopia.flow.test.parse
 
 import utopia.flow.generic.DataType
-import utopia.flow.parse.XmlReader
+import utopia.flow.generic.ValueConversions._
+import utopia.flow.parse.{XmlElement, XmlReader}
+import utopia.flow.time.Now
 import utopia.flow.util.FileExtensions._
 
 /**
@@ -15,4 +17,22 @@ object XmlReadTest extends App
 	
 	val xml = XmlReader.parseFile("Flow/data/test-data/example.xml").get
 	println(xml.toXml)
+	
+	val xml2 = xml.mutableCopy()
+	val order = xml2/"FuelFlightLeg"/"Fuel"/"FuelOrder"
+	order.name.namespace.use { implicit ns =>
+		order += XmlElement("Acknowledgement", children = Vector(
+			XmlElement("DateTime", Now.toValue),
+			XmlElement("Acknowledged", true)
+		))
+	}
+	println(xml2.toXml)
+	
+	val mapped = xml.mapPath("FuelFlightLeg", "Fuel", "FuelOrder") { order =>
+		order + XmlElement("Acknowledgement", children = Vector(
+			XmlElement("DateTime", Now.toValue),
+			XmlElement("Acknowledged", true)
+		))
+	}
+	println(mapped.toXml)
 }